refactor: Move `*Namespace`, `*Concat`
