+++ narwhals/_plan/arrow/namespace.py
@@ -9,8 +9,8 @@ import pyarrow.compute as pc  # ignore-banned-import
 from narwhals._arrow.utils import narwhals_to_native_dtype
 from narwhals._plan._guards import is_tuple_of
 from narwhals._plan.arrow import functions as fn
+from narwhals._plan.compliant.namespace import EagerNamespace
 from narwhals._plan.expressions.literal import is_literal_scalar
-from narwhals._plan.protocols import EagerNamespace
 from narwhals._utils import Version
 from narwhals.exceptions import InvalidOperationError
