+++ narwhals/_plan/compliant/typing.py
@@ -7,12 +7,12 @@ from narwhals._typing_compat import TypeVar
 if TYPE_CHECKING:
     from typing_extensions import TypeAlias
 
+    from narwhals._plan.compliant.namespace import CompliantNamespace
     from narwhals._plan.compliant.series import CompliantSeries
     from narwhals._plan.protocols import (
         CompliantBaseFrame,
         CompliantDataFrame,
         CompliantExpr,
-        CompliantNamespace,
         CompliantScalar,
         EagerDataFrame,
         EagerExpr,
